+++ src/scripts/services/dimItemService.factory.js
@@ -4,10 +4,16 @@ import _ from 'underscore';
 angular.module('dimApp')
   .factory('dimItemService', ItemService);
 
-function ItemService(dimStoreService,
-                     dimBungieService,
-                     $q,
-                     $i18next) {
+/**
+ * A service for moving/equipping items. dimItemMoveService should be preferred for most usages.
+ */
+function ItemService(
+  dimStoreService,
+  ItemFactory,
+  Destiny1Api,
+  $q,
+  $i18next
+) {
   // We'll reload the stores to check if things have been
   // thrown away or moved and we just don't have up to date info. But let's
   // throttle these calls so we don't just keep refreshing over and over.
@@ -17,18 +23,10 @@ function ItemService(dimStoreService,
   }, 10000, { trailing: false });
 
   return {
-    getSimilarItem: getSimilarItem,
-    getItem: getItem,
-    getItems: getItems,
-    moveTo: moveTo,
-    equipItems: equipItems,
-    setItemState: setItemState
+    getSimilarItem,
+    moveTo,
+    equipItems
   };
-
-  function setItemState(item, store, lockState, type) {
-    return dimBungieService.setItemState(item, store, lockState, type)
-      .then(() => lockState);
-  }
 
   /**
    * Update our item and store models after an item has been moved (or equipped/dequipped).
@@ -38,7 +36,7 @@ function ItemService(dimStoreService,
     // Refresh all the items - they may have been reloaded!
     source = dimStoreService.getStore(source.id);
     target = dimStoreService.getStore(target.id);
-    item = getItem(item);
+    item = dimStoreService.getItemAcrossStores(item);
 
     // If we've moved to a new place
     if (source.id !== target.id) {
@@ -89,7 +87,7 @@ function ItemService(dimStoreService,
           source.removeItem(sourceItem);
           sourceItem = angular.copy(sourceItem);
           sourceItem.amount -= amountToRemove;
-          sourceItem.index = dimStoreService.createItemIndex(sourceItem);
+          sourceItem.index = ItemFactory.createItemIndex(sourceItem);
           source.addItem(sourceItem);
         }
 
@@ -105,7 +103,7 @@ function ItemService(dimStoreService,
           targetItem = item;
           if (!removedSourceItem) {
             targetItem = angular.copy(item);
-            targetItem.index = dimStoreService.createItemIndex(targetItem);
+            targetItem.index = ItemFactory.createItemIndex(targetItem);
           }
           removedSourceItem = false; // only move without cloning once
           targetItem.amount = 0; // We'll increment amount below
@@ -121,7 +119,7 @@ function ItemService(dimStoreService,
         target.removeItem(targetItem);
         targetItem = angular.copy(targetItem);
         targetItem.amount += amountToAdd;
-        targetItem.index = dimStoreService.createItemIndex(targetItem);
+        targetItem.index = ItemFactory.createItemIndex(targetItem);
         target.addItem(targetItem);
         addAmount -= amountToAdd;
       }
@@ -252,7 +250,7 @@ function ItemService(dimStoreService,
       if (items.length === 1) {
         return equipItem(items[0]);
       }
-      return dimBungieService.equipItems(store, items)
+      return Destiny1Api.equipItems(store, items)
         .then((equippedItems) => {
           return equippedItems.map((i) => {
             return updateItemModel(i, store, store, true);
@@ -265,7 +263,7 @@ function ItemService(dimStoreService,
     if ($featureFlags.debugMoves) {
       console.log('Equip', item.name, item.type, 'to', dimStoreService.getStore(item.owner).name);
     }
-    return dimBungieService.equip(item)
+    return Destiny1Api.equip(item)
       .then(() => {
         const store = dimStoreService.getStore(item.owner);
         return updateItemModel(item, store, store, true);
@@ -298,7 +296,7 @@ function ItemService(dimStoreService,
     if ($featureFlags.debugMoves) {
       console.log('Move', amount, item.name, item.type, 'to', store.name, 'from', dimStoreService.getStore(item.owner).name);
     }
-    return dimBungieService.transfer(item, store, amount)
+    return Destiny1Api.transfer(item, store, amount)
       .then(() => {
         const source = dimStoreService.getStore(item.owner);
         const newItem = updateItemModel(item, source, store, false, amount);
@@ -724,18 +722,5 @@ function ItemService(dimStoreService,
 
         return promise;
       });
-  }
-
-  function getItems() {
-    let returnValue = [];
-    dimStoreService.getStores().forEach((store) => {
-      returnValue = returnValue.concat(store.items);
-    });
-    return returnValue;
-  }
-
-  function getItem(params, store) {
-    const items = store ? store.items : getItems();
-    return _.findWhere(items, _.pick(params, 'id', 'hash', 'notransfer'));
   }
 }