+++ src/scripts/shell/dimSettingsCtrl.controller.js
@@ -3,7 +3,7 @@ import _ from 'underscore';
 
 angular.module('dimApp').controller('dimSettingsCtrl', SettingsController);
 
-function SettingsController(loadingTracker, dimSettingsService, $scope, dimCsvService, dimStoreService, dimInfoService, OAuthTokenService, $state) {
+function SettingsController(loadingTracker, dimSettingsService, $scope, dimCsvService, dimStoreService, dimInfoService, OAuthTokenService, $state, $i18next) {
   const vm = this;
 
   vm.featureFlags = {
@@ -17,9 +17,9 @@ function SettingsController(loadingTracker, dimSettingsService, $scope, dimCsvSe
     dimSettingsService.save();
   });
 
-  vm.charColOptions = _.range(3, 6).map((num) => ({ id: num, name: num }));
-  vm.vaultColOptions = _.range(5, 21).map((num) => ({ id: num, name: num }));
-  vm.vaultColOptions.unshift({ id: 999, name: 'Auto' });
+  vm.charColOptions = _.range(3, 6).map((num) => ({ id: num, name: $i18next.t('Settings.ColumnSize', { num }) }));
+  vm.vaultColOptions = _.range(5, 21).map((num) => ({ id: num, name: $i18next.t('Settings.ColumnSize', { num }) }));
+  vm.vaultColOptions.unshift({ id: 999, name: $i18next.t('Settings.ColumnSizeAuto') });
 
   vm.languageOptions = {
     de: 'Deutsch',